refactor: make compliant with new export + remove API funcs
The timechimp API was never usable for public use, i just managed to find a workaround at the time.
They have moved to a new API and I couldnt find a way to use it here.
So API functionality is removed. Users should export to CSV, format it to JSON and use it that way.